refactor: Profile Response To Core Module
