+++ data/profile/src/main/java/com/apeun/gidaechi/data/profile/ProfileRepository.kt
@@ -1,7 +1,7 @@
 package com.apeun.gidaechi.data.profile
 
 import com.apeun.gidaechi.common.model.Result
-import com.apeun.gidaechi.data.profile.model.ProfileModel
+import com.apeun.gidaechi.data.core.model.ProfileModel
 import kotlinx.coroutines.flow.Flow
 
 interface ProfileRepository {
