--- data/profile/src/main/java/com/apeun/gidaechi/data/profile/mapper/ProfileMapper.kt
@@ -1,15 +1,0 @@
-package com.apeun.gidaechi.data.profile.mapper
-
-import com.apeun.gidaechi.data.profile.model.ProfileModel
-import com.apeun.gidaechi.network.profile.response.ProfileResponse
-
-internal fun ProfileResponse.toModel() = ProfileModel(
-    status = status,
-    memberId = memberId,
-    nick = nick,
-    spot = spot,
-    belong = belong,
-    phone = phone,
-    wire = wire,
-    location = location,
-)
