+++ network/profile/src/main/java/com/apeun/gidaechi/network/profile/ProfileDataSource.kt
@@ -1,7 +1,7 @@
 package com.apeun.gidaechi.network.profile
 
 import com.apeun.gidaechi.network.core.response.BaseResponse
-import com.apeun.gidaechi.network.profile.response.ProfileResponse
+import com.apeun.gidaechi.network.core.response.ProfileResponse
 
 interface ProfileDataSource {
     suspend fun loadUserInfo(workspaceId: String): BaseResponse<ProfileResponse>
