+++ network/profile/src/main/java/com/apeun/gidaechi/network/profile/datasource/ProfileDataSourceImpl.kt
@@ -3,9 +3,9 @@ package com.apeun.gidaechi.network.profile.datasource
 import com.apeun.gidaechi.network.core.SeugiUrl
 import com.apeun.gidaechi.network.core.Test
 import com.apeun.gidaechi.network.core.response.BaseResponse
+import com.apeun.gidaechi.network.core.response.ProfileResponse
 import com.apeun.gidaechi.network.core.utiles.addTestHeader
 import com.apeun.gidaechi.network.profile.ProfileDataSource
-import com.apeun.gidaechi.network.profile.response.ProfileResponse
 import io.ktor.client.HttpClient
 import io.ktor.client.call.body
 import io.ktor.client.request.get
